// commerceos/libs/apps/products/src/lib/shared/enums/product.enum.ts
export enum ProductEditorSections {
  Main = 'main',
  Pricing = 'pricing',
  Content = 'content',
  Inventory = 'inventory',
  Category = 'category',
  Variants = 'variants',
  Attributes = 'attributes',
  Channels = 'channels',
  Shipping = 'shipping',
  Taxes = 'taxes',
  Recommendations = 'recommendations',
  // RecurringBilling = 'recurring_billing',
  Visibility = 'visibility',
  SEO = 'seo',
}

export enum ProductTypes {
  Digital = 'digital',
  Physical = 'physical',
  Service = 'service',
}

export enum ProductConditions {
  New = 'new',
  Refurbished = 'refurbished',
  Used = 'used',
}

export enum ChannelTypes {
  Pos = 'pos',
  Shop = 'shop',
  Market = 'market',
  Facebook = 'facebook',
  Mobilede = 'mobilede',
  Ebay = 'ebay',
  ExternalInventory = 'external_inventory',
  Autoscout24 = 'autoscout24',
  Amazon = 'amazon',
  Marketing = 'marketing',
  Instagram = 'instagram',
  FinanceExpress = 'finance_express',
  Link = 'link',
  Dropshipping = 'dropshipping',
  GoogleShopping = 'google_shopping',
  ShopifyProduct = 'shopify-product',
  Woocommerce = 'woo_commerce',
  Plentymarkets = 'plentymarkets',
  DanDomain = 'dandomain',
  xtCommerce = 'xt_commerce',
  OXID = 'oxid',
  JTL = 'jtl',
  Shopware = 'shopware',
  PrestaShop = 'presta',
  Magento = 'magento',
  Commercetools = 'commercetools',
  Opencart = 'opencart',
  Kaufland = 'kaufland',
}