// commerceos/libs/apps/products/src/lib/shared/interfaces/section.interface.ts
export interface BaseSection {
  images?: string[];
  price: number;
  salePrice: number;
  onSales: boolean;
  available: boolean;
  productType: ProductTypes;
  saleStartDate?: string;
  saleEndDate?: string;
}

export interface MainSection extends BaseSection {
  title: string;
  condition: ProductConditions;
  brand: string;
}
// ...
export interface InventorySection {
  sku: string;
  barcode: string;
  inventory: number;
  lowInventory?: number;
  emailLowStock?: boolean;
  inventoryTrackingEnabled: boolean;
}
// ...
export interface CategorySection {
  categories: Array<Category | NotFullCategory>; // after click Save will come as Category
  channelSetCategories?: ChannelSetCategoriesInterface[];
}
// ...
export interface Category {
  id: string;
  slug: string;
  title: string;
  businessUuid: string;
}

export interface NotFullCategory {
  title: string;
}

export interface ExternalError {
  section: string;
  field: MainSection | InventorySection | string;
  errorText: string;
}